requirements.txt
streamlit
spotipy
requests
python-dotenv
